Handling all buildable types

+++ game_engine_dev/misc/bit_array.cpp
@@ -9,9 +9,9 @@
 //=> - Global constants -
 //================================================================================================================================
 
-const int MAX_BITS = 32;
+const u32 MAX_BITS = 32;
 
-constexpr uint32_t BitArray32::masks[MAX_BITS] = {
+constexpr u32 BitArray32::masks[MAX_BITS] = {
     0x00000001, 0x00000002, 0x00000004, 0x00000008,
     0x00000010, 0x00000020, 0x00000040, 0x00000080,
     0x00000100, 0x00000200, 0x00000400, 0x00000800,
@@ -26,25 +26,25 @@ constexpr uint32_t BitArray32::masks[MAX_BITS] = {
 //=> - BitArray32 implementation -
 //================================================================================================================================
 
-int BitArray32::get_count () const {
+u32 BitArray32::get_count () const {
     return m_num_bits;
 }
 
-int BitArray32::get_bit (int index) const {
+u32 BitArray32::get_bit (u32 index) const {
     if (index < 0 || index >= MAX_BITS) {
         return 0;
     }
     return (m_num_bits & masks[index]) ? 1 : 0;
 }
 
-void BitArray32::set_bit (int index) {
+void BitArray32::set_bit (u32 index) {
     if (index < 0 || index >= MAX_BITS) {
         return;
     }
     m_num_bits |= masks[index];
 }
 
-void BitArray32::clear_bit (int index) {
+void BitArray32::clear_bit (u32 index) {
     if (index < 0 || index >= MAX_BITS) {
         return;
     }
@@ -52,12 +52,12 @@ void BitArray32::clear_bit (int index) {
 }
 
 void BitArray32::serialize (std::ostream& os) const {
-    os.write(reinterpret_cast<const char*>(&m_num_bits), sizeof(uint32_t));
+    os.write(reinterpret_cast<const char*>(&m_num_bits), sizeof(u32));
 }
 
 BitArray32 BitArray32::deserialize (std::istream& is) {
-    uint32_t bits;
-    is.read(reinterpret_cast<char*>(&bits), sizeof(uint32_t));
+    u32 bits;
+    is.read(reinterpret_cast<char*>(&bits), sizeof(u32));
     return BitArray32(bits);
 }
 
@@ -65,7 +65,7 @@ BitArray32 BitArray32::deserialize (std::istream& is) {
 //=> - BitArrayCL implementation -
 //================================================================================================================================
 
-BitArrayCL::BitArrayCL (uint32_t num_bits) : m_num_bits(num_bits) {
+BitArrayCL::BitArrayCL (u32 num_bits) : m_num_bits(num_bits) {
     m_num_arrays = (num_bits + MAX_BITS - 1) / MAX_BITS;
     m_arrays = new BitArray32[m_num_arrays];
 }
@@ -74,50 +74,50 @@ BitArrayCL::~BitArrayCL () {
     delete[] m_arrays;
 }
 
-int BitArrayCL::get_count () const {
+u32 BitArrayCL::get_count () const {
     return m_num_bits;
 }
 
 
-int BitArrayCL::get_bit (int index) const {
-    if (index < 0 || index >= static_cast<int>(m_num_bits)) {
+u32 BitArrayCL::get_bit (u32 index) const {
+    if (index < 0 || index >= m_num_bits) {
         return 0;
     }
-    int array_idx = index / MAX_BITS;
-    int bit_idx = index % MAX_BITS;
+    u32 array_idx = index / MAX_BITS;
+    u32 bit_idx = index % MAX_BITS;
     return m_arrays[array_idx].get_bit(bit_idx);
 }
 
-void BitArrayCL::set_bit (int index) {
-    if (index < 0 || index >= static_cast<int>(m_num_bits)) {
+void BitArrayCL::set_bit (u32 index) {
+    if (index < 0 || index >= m_num_bits) {
         return;
     }
-    int array_idx = index / MAX_BITS;
-    int bit_idx = index % MAX_BITS;
+    u32 array_idx = index / MAX_BITS;
+    u32 bit_idx = index % MAX_BITS;
     m_arrays[array_idx].set_bit(bit_idx);
 }
 
-void BitArrayCL::clear_bit (int index) {
-    if (index < 0 || index >= static_cast<int>(m_num_bits)) {
+void BitArrayCL::clear_bit (u32 index) {
+    if (index < 0 || index >= m_num_bits) {
         return;
     }
-    int array_idx = index / MAX_BITS;
-    int bit_idx = index % MAX_BITS;
+    u32 array_idx = index / MAX_BITS;
+    u32 bit_idx = index % MAX_BITS;
     m_arrays[array_idx].clear_bit(bit_idx);
 }
 
 void BitArrayCL::serialize (std::ostream& os) const {
-    os.write(reinterpret_cast<const char*>(&m_num_bits), sizeof(uint32_t));
-    for (uint32_t i = 0; i < m_num_arrays; i++) {
+    os.write(reinterpret_cast<const char*>(&m_num_bits), sizeof(u32));
+    for (u32 i = 0; i < m_num_arrays; i++) {
         m_arrays[i].serialize(os);
     }
 }
 
 BitArrayCL* BitArrayCL::deserialize (std::istream& is) {
-    uint32_t num_bits;
-    is.read(reinterpret_cast<char*>(&num_bits), sizeof(uint32_t));
+    u32 num_bits;
+    is.read(reinterpret_cast<char*>(&num_bits), sizeof(u32));
     BitArrayCL* ba = new BitArrayCL(num_bits);
-    for (uint32_t i = 0; i < ba->m_num_arrays; i++) {
+    for (u32 i = 0; i < ba->m_num_arrays; i++) {
         ba->m_arrays[i] = BitArray32::deserialize(is);
     }
     return ba;
